Improve completion by delegating to Click
Previously we generated our own completions for available options and
commands, in a similar way to how this was is done in the
`click._bashcomplete` module.

However, this module has had various improvements made to it since this
was originally done. Rather than duplicate similar code here (and have
to do this again in future), this commit changes our completion to
delegate generating the available completions to Click. There is also
additional handling around this to still display relevant help alongside
the completions, where possible.

The end result is that this commit should make click-repl's completion
function as before, except with some additional completions shown where
these are relevant, such as the completion of choices added in
https://github.com/pallets/click/pull/681. Any additional improvements
to Click's completion will also become available to us.