Reduce duplicated option completions
Frequently a CLI has options with both short and long identifiers, e.g.
`-o` and `--opt`. For an interactive CLI such as that generated by
click-repl I think it is nicer to just show the long option(s) in these
cases, to avoid cluttering the completions with the same option many
times in different forms. This commit does this; note that short options
are still shown if there is not a corresponding long option to show
instead.