add cli call to README examples
I took the example code as "how you run a click-repl app", and was confused when the examples didn't work at all.  I figured it out for the 'advanced usage' example, but when I read the initial example of how to use `register_repl` I assumed that that replaced the `cli()` call.

Of course, it doesn't.  The fact that `register_repl` only _registers the repl_ command is implied by the name, but for such a small example snippet, where everything else needed to function is included, not including the `cli()` line implied to me that it wasn't necessary.  

This update just makes the `cli()` call explicit in both examples, and hopefully nobody else will suffer from my confusion.

Related issue: #44